refactor(app_runtime): wizard open helpers を wizard.rs に追加 (SPEC-2077 Phase F2a-1)
- crates/gwt/src/app_runtime/wizard.rs に open_launch_wizard / open_launch_wizard_for_branch / refresh_open_launch_wizard_from_cache の 3 method を追加 (Phase F1 の state helpers の隣)
- mod.rs から該当 method を削除
- main.rs の使われなくなった LaunchWizardContext import を削除
- 残: F2a-2 (issue 系 open) / F2b (handle_launch_wizard_action ~600 行) / F3a/F3b (spawn_wizard_shell_window*) は別 PR

Verification:
- cargo test -p gwt-core -p gwt: 全 pass
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
- mod.rs から ~111 行削除、 wizard.rs +112 行
